==> scripts/2. EDA.Rmd <==
---
title: "Разведочный анализ данных"
author: "Evgenii Berdinskikh"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(corrplot)
library(corrr)
library(conflicted)
library(forecast)
library(ggpubr)
library(tidyverse)

conflict_prefer("filter", "dplyr", "stats")
conflict_prefer("lag", "dplyr", "stats")

theme_custom <- theme_bw()+ theme(
    plot.title = element_text(size = 30, hjust = 0.5),
    plot.subtitle = element_text(size = 25, hjust = 0.5),
    strip.text = element_text(size = 20),
    axis.text = element_text(size = 20),
    axis.title = element_text(size = 25),
    legend.title = element_text(size = 25),
    legend.text = element_text(size = 20)
  )

```

## Загрузка данных для анализа

```{r read, message=FALSE}

brain_data <- read_csv("../data/raw/brain_data.csv") %>%
   mutate(across(c(Month, Year, Sud_Imp, Sud_Storm), as.factor),
          Dst_level = factor(Dst_level, 
                             levels = c("Quiet", "Weak", "Moderate", "Major", "Severe"),
                             ordered = TRUE))

```

## Корреляция количественных переменных

```{r cor, fig.width=10, message=FALSE}

cor_data <- brain_data %>% 
  select( where(is.numeric) & !c(Temp_min, Temp_max, Dst_PC1, Dst_PC2) ) %>%
  cor(use = "pairwise.complete.obs") 
  
corrplot(cor_data, method = 'number', type = 'lower', diag = FALSE)

network_plot(cor_data, min_cor = .1)

```

При анализе сетевого графика можно выделить три группы признаков:

- Климатические: температура, давление, ветер, осадки;
- Геомагнитные: индексы Dst, Ap;
- Солнечные: индекс F10.7, количество солнечных пятен.

При этом некоторые признаки внутри групп обладают высокой степенью корреляции!

```{r point, message=FALSE}

brain_data %>% 
  ggplot(aes(Sunspots, F10.7))+
  geom_jitter(alpha = 0.5, colour = "royalblue2", size = 1.5)+
  geom_smooth(method = "lm", se = FALSE, colour = "orangered2", linewidth = 1.2) +
  stat_cor(method = "pearson", label.x.npc = 0.4, label.y.npc = "top", size = 5)+
  theme_custom

brain_data %>% 
  ggplot(aes(Dst_mean, Ap_mean))+
  geom_jitter(alpha = 0.5, colour = "royalblue2", size = 1.5)+
  geom_smooth(method = "lm", se = FALSE, colour = "orangered2", linewidth = 1.2) +
  stat_cor(method = "pearson", label.x.npc = 0.4, label.y.npc = "top", size = 5)+
  theme_custom

brain_data %>% 
  drop_na() %>% 
  ggplot(aes(Temp_mean, Pressure))+
  geom_jitter(alpha = 0.5, colour = "royalblue2", size = 1.5)+
  geom_smooth(method = "lm", se = FALSE, colour = "orangered2", linewidth = 1.2) +
  stat_cor(method = "pearson", label.x.npc = 0.4, label.y.npc = "top", size = 5)+
  theme_custom

```

- Индекс **F10.7** и количество **солнечных пятен** показывают высокую корреляцию (r = 0.92), поскольку оба показателя отражают солнечную активность, но измеряются разными методами.  
- Среднесуточные индексы **Ap** и **Dst** также заметно коррелируют (r = -0.70), так как оба характеризуют изменения в геомагнитном поле Земли под воздействием солнечного ветра.  
- Также ожидаемо выявлена отрицательная корреляция между средней температурой и давлением за день (r = -0.73).  

Для дальнейшего анализа из каждой пары переменных оставим одну:  
1. **Количество солнечных пятен** — этот показатель имеет полный набор данных и обладает большей вариативностью.  
2. **Индекс Dst** — напрямую связан с влиянием солнечной активности на магнитосферу.  
3. **Температура** — обеспечивает полный охват данных за весь период исследования и обладает более высокой вариативностью.  

```{r cor2, fig.width=10, message=FALSE}

cor_data2 <- brain_data %>% 
  select( where(is.numeric) & !c(Temp_min, Temp_max, F10.7, Pressure, Ap_mean) ) %>%
  cor(use = "pairwise.complete.obs") 
  
corrplot(cor_data2, method = 'number', type = 'lower', diag = FALSE)

network_plot(cor_data2, min_cor = .1)

```


Влияние Dst_PC1 и Dst_PC2 на рост числа вызовов скорой помощи, вероятно, схоже с Dst_mean и Dst_var, но их стоит дополнительно проверить на этапе моделирования.


## Циклы солнечной активности

```{r cycle}

brain_data %>% 
  select(!c(Pressure, F10.7, Ap_mean)) %>% 
  ggplot(aes(Date, Sunspots))+
  geom_line(alpha = 0.3, colour = "royalblue2")+
  geom_smooth(method = "gam", formula = y ~ s(x, k = 10),  se = FALSE,
              colour = "orangered4", linewidth = 2) +
  geom_vline(xintercept = as.Date("2009-01-01"), 
             colour = "orangered", linewidth = 1, linetype = "dashed")+
  geom_vline(xintercept = as.Date("2019-12-31"), 
             colour = "orangered", linewidth = 1, linetype = "dashed")+
  scale_x_date(date_breaks = "1 year", date_labels = "%Y")+
  theme_bw()

```

Количество солнечных пятен напрямую связано с циклами солнечной активности, что подтверждается на графике. Для дальнейшего анализа целесообразно отдельно рассмотреть период 2009–2019 гг., охватывающий полный 24-й цикл солнечной активности

## Декомпозиция данных

```{r decompose, fig.height=8}

brain_data %>% 
  filter(Date >=  ymd("2007-01-01")) %>%
  pull(Sunspots) %>% ts(start = 2007, frequency = 365) %>%
  stl(s.window = "periodic") %>% 
  autoplot() + 
  theme_custom

```

Декомпозиция позволяет выявить тренды и сезонную цикличность:

```{r check, fig.width=12, message=FALSE, warning=FALSE}

##Линейный тренд EMS
brain_data %>% 
  filter(Date >=  ymd("2007-01-01")) %>%
  pull(EMS) %>% ts(start = 2007, frequency = 365) %>%
  stl(s.window = "periodic") %>% 
  .$time.series %>% .[,"trend"] %>%
  autoplot() + 
  scale_x_continuous(n.breaks = 10)+
  theme_custom+
  ggtitle("Декомпозиция вызовов скорой помощи: Тренд")


##Сезонная цикличность по температуре
brain_data %>% 
  filter(Date >=  ymd("2007-01-01")) %>%
  pull(Temp_mean) %>% ts(start = 2007, frequency = 365) %>%
  stl(s.window = "periodic") %>% 
  .$time.series %>% .[,"seasonal"] %>%
  autoplot()+
  scale_x_continuous(n.breaks = 10)+
  theme_custom+
  ggtitle("Декомпозиция температуры: Сезонность")


# Сложности с индексом Dst
ggarrange(
brain_data %>% 
  filter(Date >=  ymd("2007-01-01")) %>%
  pull(Sunspots) %>% ts(start = 2007, frequency = 365) %>%
  stl(s.window = "periodic") %>% 
  .$time.series %>% .[,"trend"] %>%
  autoplot() + 
  scale_x_continuous(n.breaks = 5)+
  theme_custom, 
  
brain_data %>% 
  filter(Date >=  ymd("2007-01-01")) %>%
  pull(Dst_var) %>% ts(start = 2007, frequency = 365) %>%
  stl(s.window = "periodic") %>% 
  .$time.series %>% .[,"trend"] %>%
  autoplot() + 
  scale_x_continuous(n.breaks = 5)+
  theme_custom, 

 labels = c("Декомпозиция солнечных пятен: Тренд", "Декомпозиция индекса Dst: Тренд") 
  ) 


```

- После декомпозиции данных о вызовах скорой помощи наблюдается линейный рост во времени. Однако начиная с 2020 года заметно снижение, которое, вероятно, связано с пандемией COVID-19.
- Климатические данные (особенно температура), после декомпозиции, демонстрируют выраженную цикличность.
- Индекс Dst_mean не показывает явных трендов или циклов. В то время как индекс Dst_var (амплитуда колебаний индекса Dst за день), после декомпозиции, имеет тренд, схожий с динамикой количества солнечных пятен.

## Визуализация ассоциаций

```{r vis, fig.width=15}

#EMS/Sunspots сезонность
ggarrange(
brain_data %>%  
  filter(Date %within% interval(ymd("2007-01-01"), ymd("2019-12-31"))) %>%
  group_by(Year, Month) %>% 
  summarise(EMS_month = sum(EMS, na.rm = TRUE), .groups = "drop") %>% 
  ggplot(aes(x = Month, y = Year, fill = EMS_month)) +
  geom_tile(colour = "black") + 
  scale_fill_gradient(low = "green", high = "red") + 
  scale_y_discrete(expand = expansion(add = c(0, 1)))+
  labs(x = "", y  = "", fill = "")+
  theme_custom,

brain_data %>%  
  filter(Date %within% interval(ymd("2007-01-01"), ymd("2019-12-31"))) %>%
  group_by(Year, Month) %>% 
  summarise(Sunspots_month = sum(Sunspots, na.rm = TRUE), .groups = "drop") %>% 
  ggplot(aes(x = Month, y = Year, fill = Sunspots_month)) +
  geom_tile(colour = "black") + 
  scale_fill_gradient(low = "green", high = "red") + 
  scale_y_discrete(expand = expansion(add = c(0, 1)))+
  labs(x = "", y  = "", fill = "")+
  theme_custom, 
 
labels = c("Суммарное количество вызовов скорой", "Суммарное количество солнечных пятен")
)

#Dst ~ Sunspots
brain_data %>%
  filter(Dst_level != "Severe") %>%
  ggplot(aes(Dst_level, Sunspots))+
  geom_jitter(alpha = 0.5)+
  geom_boxplot(alpha = 0.5, outliers = FALSE, fill = "goldenrod2", linewidth = 1)+
  theme_custom

```
